refactor: 캐싱 로직 GetAllFirst1000UseCase로 옮김
