+++ app/src/main/java/com/example/seoulpublicservice/ui/notifications/NotificationsViewModel.kt
@@ -9,8 +9,6 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import com.example.seoulpublicservice.SeoulPublicServiceApplication
-import com.example.seoulpublicservice.pref.PrefRepository
-import com.example.seoulpublicservice.pref.RowPrefRepository
 import com.example.seoulpublicservice.seoul.Row
 import com.example.seoulpublicservice.usecase.GetAllFirst1000UseCase
 import kotlinx.coroutines.Dispatchers
@@ -18,9 +16,7 @@ import kotlinx.coroutines.launch
 import kotlin.random.Random
 
 class NotificationsViewModel(
-    private val getAllFirst1000UseCase: GetAllFirst1000UseCase,
-    private val rowPrefRepository: RowPrefRepository,
-    private val prefRepository: PrefRepository
+    private val getAllFirst1000UseCase: GetAllFirst1000UseCase
 ) : ViewModel() {
 
     private val _text: MutableLiveData<String> =
@@ -30,29 +26,9 @@ class NotificationsViewModel(
     private var rowList: List<Row> = emptyList()
     private val random = Random
 
-    private val tempKeyRowsSavedTime = "tempKeyRowsSavedTime"
-
     fun setRandomOne() {
         if (rowList.isEmpty()) viewModelScope.launch(Dispatchers.IO) {
-            var isRecent = false
-            val rowsSavedTime = prefRepository.load(tempKeyRowsSavedTime).toLongOrNull()
-            if (rowsSavedTime == null) {
-                Log.w(
-                    "jj-노티뷰모델",
-                    "prefRepository.load(tempKeyRowsSavedTime).toLongOrNull() == null"
-                )
-            } else {
-                val timeDiff = System.currentTimeMillis() - rowsSavedTime
-                Log.d("jj-노티뷰모델", "timeDiff: $timeDiff")
-                isRecent = timeDiff < 180_000L
-            }
-
-            if (isRecent) rowList = rowPrefRepository.loadRows()
-            if (rowList.isEmpty()) {
-                rowList = getAllFirst1000UseCase()
-                rowPrefRepository.saveRows(rowList)
-                prefRepository.save(tempKeyRowsSavedTime, System.currentTimeMillis().toString())
-            }
+            rowList = getAllFirst1000UseCase()
 
             val row = rowList.firstOrNull()
             if (row == null) {
@@ -77,9 +53,7 @@ class NotificationsViewModel(
                 val application = (this[APPLICATION_KEY] as SeoulPublicServiceApplication)
                 val container = application.container
                 NotificationsViewModel(
-                    getAllFirst1000UseCase = container.getAllFirst1000UseCase,
-                    rowPrefRepository = container.rowPrefRepository,
-                    prefRepository = container.prefRepository
+                    getAllFirst1000UseCase = container.getAllFirst1000UseCase
                 )
             }
         }
